Implement Sorting for ActivityPojo
ActivityPojos are now explicitly sorted by start time using the
compareTo() method before being displayed in LogActivity. This change
was made to accomodate an upcoming feature which would allow users to
modify details of their recorded activities (such as start time). With
this change, activities will still be displayed in chronological order
even if the database's chronological integrity is lost.